PrintFact
The Java program in the given code defines a class named `PrintFactorial` in the package `com.ritesh`. The program's purpose is to calculate and print the factorial of a given positive integer `n` using a recursive approach.

The program consists of two methods:

1. `calfactoriaL(int n)`: This method takes an integer `n` as a parameter and calculates the factorial of `n` using recursion. If `n` is either 1 or 0, it returns 1 (since the factorial of 0 and 1 is 1). Otherwise, it recursively calculates the factorial of `n-1` and then multiplies it by `n` to get the factorial of `n`.

2. `main(String[] args)`: This is the entry point of the program. It initializes the value of `n` to 5 and then calls the `calfactoriaL` method to calculate the factorial of `n`. The result is stored in the `ans` variable, which is then printed to the console.

In summary, the program demonstrates the use of recursion to compute the factorial of a given integer and outputs the result.